Part 3 of static analysis: clang-tidy (#1460)
Last part of https://github.com/LDMX-Software/ldmx-sw/issues/1176
* Fix warning from clang-tidy
* Update justfile with new compile settings

+++ TrigScint/src/TrigScint/SimQIE.cxx
@@ -43,8 +43,7 @@ int SimQIE::Q2ADC(float Charge) {
     } else
       b = (a + b) / 2;
   }
-  return 64 * (int)(a / 4) + nbins_[a % 4] +
-         floor((qq - edges_[a]) / sense_[a]);
+  return 64 * (a / 4) + nbins_[a % 4] + floor((qq - edges_[a]) / sense_[a]);
 }
 
 // Function to convert ADCs back to charge
